[WIP] Add View Cube
This PR begins adding a View Cube at the request of #26.

After this quick and dirty implementation, I'm not totally convinced the additional complexity and state are worth it.  The camera movement functionality isn't necessary in a SCAD app, given the drawing occurs entirely from the code-editor and not the graphical editor.

I may downgrade this to a simple orthonormal basis (to provide a quick reference for which direction is which).